Aggiornato IdFattura a long? per supporto bigint DB PF-672
Modifica IdFattura da int? a long? in GestioneFattureAzioneCommand, GestioneFattureAzioneRequest e GestioneFattureAzioneCommandPersistence per allineamento al tipo bigint del database. Aggiornato anche il parametro Dapper a DbType.Int64.

+++ src/Infrastructure/PortaleFatture.BE.Infrastructure/Common/SEND/Fatture/Commands/GestioneFattureAzioneCommand.cs
@@ -18,7 +18,7 @@ namespace PortaleFatture.BE.Infrastructure.Common.SEND.Fatture.Commands;
     public string? IdUtente { get; set; }
     public NoteCommand? Nota { get; set; }
 
-    public int? IdFattura { get; set; }
+    public long? IdFattura { get; set; }   // bigint a DB
 }
 
 public class NoteCommand
